% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{generate_analysis_hash}
\alias{generate_analysis_hash}
\title{Generate Analysis Report Hash}
\usage{
generate_analysis_hash(desandbox_object, parameters)
}
\arguments{
\item{desandbox_object}{The input data object}

\item{parameters}{List of analysis parameters}
}
\value{
A hash string
}
\description{
Creates a reproducible hash for tracking analysis parameters.
}
\examples{
# Internal use
}
