% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{list_cache_files}
\alias{list_cache_files}
\title{List Available Cache Files}
\usage{
list_cache_files(cache_dir = "./desandbox_cache")
}
\arguments{
\item{cache_dir}{Directory containing cache files (default: "./desandbox_cache")}
}
\value{
A data.frame with cache file information
}
\description{
Lists all cache files in the specified directory.
}
\examples{
list_cache_files()
}
