% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Global Variables and Imports}
\description{
This file defines global variables used in NSE (non-standard evaluation)
and imports commonly used functions from base R packages.
}
\keyword{internal}
