% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{run_all_methods}
\alias{run_all_methods}
\title{Run All DE Methods}
\usage{
run_all_methods(
  desandbox_object,
  methods = c("DESeq2", "edgeR", "limma-voom"),
  ...
)
}
\arguments{
\item{desandbox_object}{A SummarizedExperiment object from create_desandbox_object()}

\item{methods}{Vector of methods to run: "DESeq2", "edgeR", "limma-voom" (default: all)}

\item{...}{Additional arguments passed to individual method functions}
}
\value{
A list containing results from all methods
}
\description{
Convenience function to run DESeq2, edgeR, and limma-voom in parallel.
}
\examples{
# See vignette for complete example
}
